Move ElementHasher to its own mod
